+++ packages/common/src/project/index.ts
@@ -7,3 +7,4 @@ export * from './versioned';
 export * from './readers';
 export * from './database';
 export * from './utils';
+export * from './models';
